Add event listener to checkbox
